moved inner class ErrorResponse to package internal to remove reference to outer class.
