+++ src/main/java/me/figo/FigoException.java
@@ -22,6 +22,8 @@
 
 package me.figo;
 
+import me.figo.internal.ErrorResponse;
+
 /***
  * Base Class for all figo Exceptions. It extends the normal Java exceptions with an error_code field, which carries the computer readable error reason.
  * 
@@ -51,21 +53,5 @@ public class FigoException extends Exception {
 
     public String getErrorCode() {
         return error_code;
-    }
-
-    public class ErrorResponse {
-        private String error;
-        private String error_description;
-
-        public ErrorResponse() {
-        }
-
-        public String getError() {
-            return error;
-        }
-
-        public String getErrorDescription() {
-            return error_description;
-        }
     }
 }
